move fake cloud api callbacks to own module
